+++ pkg/bgp/mp-nlri.go
@@ -5,6 +5,7 @@ import (
 	"github.com/sbezverk/gobmp/pkg/evpn"
 	"github.com/sbezverk/gobmp/pkg/flowspec"
 	"github.com/sbezverk/gobmp/pkg/ls"
+	"github.com/sbezverk/gobmp/pkg/mcastvpn"
 	"github.com/sbezverk/gobmp/pkg/srpolicy"
 )
 
@@ -18,6 +19,7 @@ type MPNLRI interface {
 	GetNLRI71() (*ls.NLRI71, error)
 	GetNLRI73() (*srpolicy.NLRI73, error)
 	GetFlowspecNLRI() (*flowspec.NLRI, error)
+	GetNLRIMCASTVPN() (*mcastvpn.Route, error)
 	GetNextHop() string
 	IsIPv6NLRI() bool
 	IsNextHopIPv6() bool
@@ -72,6 +74,12 @@ func NLRIMessageType(afi uint16, safi uint8) int {
 		// AFI 2 and SAFI 134 FlowSpec VPNv6
 	case afi == 2 && safi == 134:
 		return 27
+		// AFI 1 and SAFI 5 MCAST-VPN v4
+	case afi == 1 && safi == 5:
+		return 32
+		// AFI 2 and SAFI 5 MCAST-VPN v6
+	case afi == 2 && safi == 5:
+		return 33
 	}
 
 	return 0